chore: move form-distribution component
